Create separated file to get dependencies of an object
